Develop

+++ db/seeds.rb
@@ -4,6 +4,8 @@
 #                                     Or run with 2 commands: rails db:drop and rails db:reset                                      #
 #***********************************************************************************************************************************#
 
+abort "only run seeds in development" unless Rails.env.development?
+
 require 'faker'
 
 Company.destroy_all
@@ -17,25 +19,34 @@ puts "Company created!"
 
 20.times do
   fake_password = "0123456789"
-  user = User.create(first_name: Faker::Name.first_name, last_name: Faker::Name.last_name, email: Faker::Internet.email, password: fake_password, password_confirmation: fake_password, company: Company.all.sample, is_validated: true, confirmed_at: Time.now)
+  user = User.create(first_name: Faker::Name.first_name, last_name: Faker::Name.last_name, email: "#{rand(1000000000)}@example.com", password: fake_password, password_confirmation: fake_password, is_validated: true, confirmed_at: Time.now)
+  user.update(company: Company.all.sample)
 end
 puts "20 users created!"
 
 #One fake site_admin_user created
 fake_password = "0123456789"
-site_admin = User.create(first_name: Faker::Name.first_name, last_name: Faker::Name.last_name, email: "site_admin@yopmail.com", password: fake_password, password_confirmation: fake_password, is_site_admin: true, is_validated: true,  confirmed_at: Time.now)
+site_admin = User.create(first_name: "Site".first_name, last_name: "Admin", email: "site_admin@example.com", password: fake_password, password_confirmation: fake_password, is_site_admin: true, is_validated: true,  confirmed_at: Time.now)
+site_admin.update(company: Company.all.first)
 puts "Site admin created!"
 
 #One fake company_admin_user created for each company
-company_admin_1 = User.create(first_name: Faker::Name.first_name, last_name: Faker::Name.last_name, email: "company_admin_1@yopmail.com", password: fake_password, password_confirmation: fake_password, company_id: 1, is_company_admin: true, is_validated: true,  confirmed_at: Time.now)
-company_admin_2 = User.create(first_name: Faker::Name.first_name, last_name: Faker::Name.last_name, email: "company_admin_2@yopmail.com", password: fake_password, password_confirmation: fake_password, company_id: 2, is_company_admin: true, is_validated: true,  confirmed_at: Time.now)
-company_admin_3 = User.create(first_name: Faker::Name.first_name, last_name: Faker::Name.last_name, email: "company_admin_3@yopmail.com", password: fake_password, password_confirmation: fake_password, company_id: 3, is_company_admin: true, is_validated: true,  confirmed_at: Time.now)
+fake_password = "0123456789"
+company_admin_1 = User.create(first_name: Faker::Name.first_name, last_name: Faker::Name.last_name, email: "company_admin_1@example.com", password: fake_password, password_confirmation: fake_password, is_company_admin: true, is_validated: true,  confirmed_at: Time.now)
+company_admin_1.update(company: Company.all.first)
+company_admin_2 = User.create(first_name: Faker::Name.first_name, last_name: Faker::Name.last_name, email: "company_admin_2@example.com", password: fake_password, password_confirmation: fake_password, is_company_admin: true, is_validated: true,  confirmed_at: Time.now)
+company_admin_2.update(company: Company.all[1])
+company_admin_3 = User.create(first_name: Faker::Name.first_name, last_name: Faker::Name.last_name, email: "company_admin_3@example.com", password: fake_password, password_confirmation: fake_password, is_company_admin: true, is_validated: true,  confirmed_at: Time.now)
+company_admin_3.update(company: Company.all[2])
 
 puts "Company admin created!"
 #One fake test_user created for each company
-test_user_1 = User.create(first_name: "Jean", last_name: "Dupont", email: "jean_dupont_1@yopmail.com", password: fake_password, password_confirmation: fake_password, company_id: 1, is_validated: true,  confirmed_at: Time.now)
-test_user_2 = User.create(first_name: "Jean", last_name: "Dupont", email: "jean_dupont_2@yopmail.com", password: fake_password, password_confirmation: fake_password, company_id: 2, is_validated: true,  confirmed_at: Time.now)
-test_user_3 = User.create(first_name: "Jean", last_name: "Dupont", email: "jean_dupont_3@yopmail;com", password: fake_password, password_confirmation: fake_password, company_id: 3, is_validated: true,  confirmed_at: Time.now)
+test_user_1 = User.create(first_name: "Jean", last_name: "Dupont", email: "jean_dupont_1@example.com", password: fake_password, password_confirmation: fake_password, company_id: 1, is_validated: true,  confirmed_at: Time.now)
+test_user_1.update(company: Company.all[0])
+test_user_2 = User.create(first_name: "Jean", last_name: "Dupont", email: "jean_dupont_2@example.com", password: fake_password, password_confirmation: fake_password, company_id: 2, is_validated: true,  confirmed_at: Time.now)
+test_user_2.update(company: Company.all[1])
+test_user_3 = User.create(first_name: "Jean", last_name: "Dupont", email: "jean_dupont_3@example.com", password: fake_password, password_confirmation: fake_password, company_id: 3, is_validated: true,  confirmed_at: Time.now)
+test_user_3.update(company: Company.all[2])
 puts "Test users created!"
 
 100.times do
